requirements.txt
streamlit
requests
azure-storage-queue
pandas
matplotlib